code cleaning

+++ src/main/java/com/github/sportbot/dto/ExerciseEntryRequest.java
@@ -4,7 +4,7 @@ import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 
 public record ExerciseEntryRequest(
-        @NotNull Long telegramId,
+        @NotNull Integer telegramId,
         @NotBlank String exerciseType,
         int count
 ) {}
